Fixes Amaze issue for VideoWidget
