+++ app/src/org/commcare/views/widgets/MediaWidget.java
@@ -1,13 +1,24 @@
 package org.commcare.views.widgets;
 
+import android.content.ActivityNotFoundException;
+import android.content.ContentValues;
 import android.content.Context;
+import android.content.Intent;
+import android.net.Uri;
+import android.provider.MediaStore;
 import android.util.Log;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
 import android.widget.LinearLayout;
+import android.widget.Toast;
 
+import org.commcare.CommCareApplication;
 import org.commcare.activities.components.FormEntryInstanceState;
+import org.commcare.dalvik.R;
 import org.commcare.logic.PendingCalloutInterface;
+import org.commcare.utils.FileUtil;
+import org.commcare.utils.StringUtils;
+import org.commcare.utils.UriToFilePath;
 import org.javarosa.core.model.data.IAnswerData;
 import org.javarosa.core.model.data.IntegerData;
 import org.javarosa.core.model.data.InvalidData;
@@ -15,6 +26,9 @@ import org.javarosa.core.model.data.StringData;
 import org.javarosa.form.api.FormEntryPrompt;
 
 import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
 
 /**
  * Generic logic for capturing or choosing audio/video/image media
@@ -23,6 +37,8 @@ import java.io.File;
  */
 public abstract class MediaWidget extends QuestionWidget {
     private static final String TAG = MediaWidget.class.getSimpleName();
+
+    protected static final String CUSTOM_TAG = "custom";
 
     protected Button mCaptureButton;
     protected Button mPlayButton;
@@ -33,6 +49,10 @@ public abstract class MediaWidget extends QuestionWidget {
     protected final String mInstanceFolder;
 
     private int oversizedMediaSize;
+
+    protected String recordedFileName;
+    protected String customFileTag;
+    protected String destMediaPath;
 
     public MediaWidget(Context context, FormEntryPrompt prompt,
                        PendingCalloutInterface pendingCalloutInterface) {
@@ -128,8 +148,6 @@ public abstract class MediaWidget extends QuestionWidget {
         mBinaryName = null;
     }
 
-    protected abstract String createFilePath(Object binaryUri);
-
     @Override
     public void setOnLongClickListener(OnLongClickListener l) {
         mCaptureButton.setOnLongClickListener(l);
@@ -161,5 +179,118 @@ public abstract class MediaWidget extends QuestionWidget {
         InputMethodManager inputManager =
                 (InputMethodManager)context.getSystemService(Context.INPUT_METHOD_SERVICE);
         inputManager.hideSoftInputFromWindow(this.getWindowToken(), 0);
+    }
+
+    @Override
+    public void setBinaryData(Object binaryuri) {
+        String binaryPath = getBinaryPathWithSizeCheck(binaryuri);
+        if (binaryPath == null) {
+            // if we have already copied the file to destMediaPath, delete it
+            if (!destMediaPath.isEmpty()) {
+                new File(destMediaPath).delete();
+            }
+            return;
+        }
+
+        File newMedia;
+
+        if (!destMediaPath.isEmpty()) {
+            // we have already copied the file at newPath during createFilePath
+            newMedia = new File(destMediaPath);
+        } else {
+            recordedFileName = FileUtil.getFileName(binaryPath);
+            destMediaPath = mInstanceFolder + System.currentTimeMillis() + customFileTag + FileUtil.getExtension(binaryPath);
+
+            // Copy to destMediaPath
+            File source = new File(binaryPath);
+            newMedia = new File(destMediaPath);
+            try {
+                FileUtil.copyFile(source, newMedia);
+            } catch (IOException e) {
+                showToast("form.attachment.copy.fail");
+                Log.e(TAG, "IOExeception while copying media");
+                e.printStackTrace();
+            }
+        }
+
+        if (newMedia.exists()) {
+            // Add the copy to the content provider
+            ContentValues values = new ContentValues(6);
+            values.put(MediaStore.Audio.Media.TITLE, newMedia.getName());
+            values.put(MediaStore.Audio.Media.DISPLAY_NAME, newMedia.getName());
+            values.put(MediaStore.Audio.Media.DATE_ADDED, System.currentTimeMillis());
+            values.put(MediaStore.Audio.Media.DATA, newMedia.getAbsolutePath());
+
+            Uri mediaUri =
+                    getContext().getContentResolver().insert(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI, values);
+            String mediaUriString = mediaUri == null ? "null" : mediaUri.toString();
+            Log.i(TAG, "Inserting media returned uri = " + mediaUriString);
+            showToast("form.attachment.success");
+        } else {
+            Log.e(TAG, "Inserting media file FAILED");
+        }
+
+        mBinaryName = newMedia.getName();
+    }
+
+    /**
+     * If file is chosen by user, the file selection intent will return an URI
+     * If file is auto-selected after recording_fragment, then the recordingfragment will provide a string file path
+     * Set value of customFileTag if the file is a recent recording from the RecordingFragment
+     */
+    protected String createFilePath(Object binaryuri) {
+        String path = "";
+        destMediaPath = "";
+        if (binaryuri instanceof Uri) {
+            try {
+                path = UriToFilePath.getPathFromUri(CommCareApplication.instance(),
+                        (Uri)binaryuri);
+            } catch (UriToFilePath.NoDataColumnForUriException e) {
+                // Need to make a copy of file using uri, so might as well copy to final destination path directly
+                InputStream inputStream;
+                try {
+                    inputStream = getContext().getContentResolver().openInputStream((Uri)binaryuri);
+                } catch (FileNotFoundException e1) {
+                    showToast("form.attachment.notfound");
+                    e1.printStackTrace();
+                    return "";
+                }
+
+                recordedFileName = FileUtil.getFileName(((Uri)binaryuri).getPath());
+                destMediaPath = mInstanceFolder + System.currentTimeMillis() + FileUtil.getExtension(((Uri)binaryuri).getPath());
+
+                try {
+                    FileUtil.copyFile(inputStream, new File(destMediaPath));
+                } catch (IOException e1) {
+                    e1.printStackTrace();
+                    showToast("form.attachment.copy.fail");
+                    return "";
+                }
+                path = destMediaPath;
+            }
+            customFileTag = "";
+        } else {
+            path = (String)binaryuri;
+            customFileTag = CUSTOM_TAG;
+        }
+        return path;
+    }
+
+    protected void playMedia(String mediaType) {
+        Intent i = new Intent("android.intent.action.VIEW");
+        File mediaFile = new File(mInstanceFolder + mBinaryName);
+        Uri mediaUri = FileUtil.getUriForExternalFile(getContext(), mediaFile);
+        i.setDataAndType(mediaUri, mediaType);
+
+        UriToFilePath.grantPermissionForUri(getContext(), i, mediaUri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
+        try {
+            getContext().startActivity(i);
+        } catch (ActivityNotFoundException e) {
+            Toast.makeText(getContext(),
+                    StringUtils.getStringSpannableRobust(getContext(),
+                            R.string.activity_not_found,
+                            "play media"),
+                    Toast.LENGTH_SHORT).show();
+        }
     }
 }
